Basic ACTUAL backend inference with NN

+++ Assets/Scripts/Implementation/Dummy/DummyAction.cs
@@ -4,7 +4,7 @@ namespace Implementation.Dummy {
     public struct DummyAction : IByteConvertible, IByteAssignable {
         public bool tap;
 
-        public void AssignFromBytes(byte[] bytes) { tap = bytes.ToInt() > 0; }
+        public void AssignFromBytes(byte[] bytes) { tap = bytes.ToFloat() > 0; }
         public byte[] ToBytes() => (tap ? 1f : 0f).ToBytes();
     }
 }
